requirements/prod.txt
Django==3.1.6
django-import-export==2.5.0
django-pandas==0.6.4
numpy==1.20.1
sklearn==0.0